requirements.txt
pyyaml
jinja2
pyeapi
python-docx
openpyxl

# Dev
pylint
flake8
coverage
coverage-badge

# pytest plugins
pytest
pytest-excel
pytest-xdist
pytest-html
pytest-json
pytest-cache
pytest-sugar
#pytest-spec
pytest-cov
